refactor: Change VideoPlayer to VideoInspector
